requirements.txt
requests
pandas
pyarrow
apache-airflow==2.9.1
